CRUD Imprumuturi

+++ Proiect/src/main/java/com/proiect/biblioteca/service/CarteService.java
@@ -2,9 +2,11 @@ package com.proiect.biblioteca.service;
 
 
 import com.proiect.biblioteca.domain.Carte;
+import com.proiect.biblioteca.domain.Imprumut;
 import com.proiect.biblioteca.dto.CarteDto;
 import com.proiect.biblioteca.repository.AutorRepository;
 import com.proiect.biblioteca.repository.CarteRepository;
+import com.proiect.biblioteca.repository.ImprumutRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
@@ -14,9 +16,11 @@ import java.util.List;
 @Service
 public class CarteService {
     private final CarteRepository carteRepository;
+    private final ImprumutRepository imprumutRepository;
 
-    public CarteService(CarteRepository carteRepository){
+    public CarteService(CarteRepository carteRepository, ImprumutRepository imprumutRepository){
         this.carteRepository = carteRepository;
+        this.imprumutRepository = imprumutRepository;
     }
 
     public List<Carte> getAll(){
@@ -36,6 +40,9 @@ public class CarteService {
     }
 
     public String delete(int id){
-        return carteRepository.delete(id);
+        List<Imprumut> imprumuturi = imprumutRepository.findByIdCarte(id);
+        if(imprumuturi.size() == 0)
+            return carteRepository.delete(id);
+        else return "Exista carti imprumutate.";
     }
 }
